+++ CHANGELOG.md
@@ -55,8 +55,15 @@ YOUR CHANGES/ADDED FEATURES HERE
 - Made it so that we can declare an uninitialized variable, without even explicitly declaring its type.
 - Fixed if statements so that if a variable is initialized in a peculiar branch, it should be in the other one as well, with the same type.
 
+## 2024-12-9
+
+- Begin work on backend part. Now literals, binop, unop and variables definition are implemented. WIP.
+- There is an issue with the way we define variables thanks to a block. Need to look into it.
+- Fixed the issue.
+
 ## TODO
 
+- Finish the basic backend for the compiler
 - Implement references
 - Write type rules
 - Add support for CONSTANT expression in array size, and any non unit type expression in the init value during array definition.